lift state up TodoForm
